Started rewriting PC code to make it better suitable for future integration into qml dashboard, improved error handling for sensor in firmware

+++ Firmware/src/Display.cpp
@@ -29,12 +29,12 @@ namespace CO2::Firmware
         m_LCD.print(buffer);
     }
 
-    void Display::DrawDashboard(const float temp, const float hum, const float co2)
+    void Display::DrawDashboard(const float temp, const float hum, const uint32_t co2)
     {
         m_LCD.clear();
 
         m_LCD.setCursor(0, 0);
-        m_LCD.printf("CO2: %.2f PPM", co2);
+        m_LCD.printf("CO2: %u PPM", co2);
 
         m_LCD.setCursor(0, 1);
         m_LCD.printf("T: %.1f C, H: %.1f", temp, hum);
