+++ Firmware/src/Sensor/Sensor.hpp
@@ -17,11 +17,11 @@ namespace CO2::Firmware
         void BeginTask();
 
         QueueHandle_t GetQueue() const { return m_Queue; }
-        void GetDisplayStats(float& temp, float& hum, float& co2);
+        void GetDisplayStats(float& temp, float& hum, uint32_t& co2);
     private:
         RunningAvgHelper<float, 5> m_TempAvg;
         RunningAvgHelper<float, 5> m_HumAvg;
-        RunningAvgHelper<float, 5> m_CO2Avg;
+        RunningAvgHelper<uint32_t, 5> m_CO2Avg;
 
         QueueHandle_t m_Queue;
         SensirionI2cScd4x m_SCD40;
